refactor: MyAccount 페이지 구조 변경 및 CompanyRecruitments로 파일명 변경
- 기존 MyAccount 페이지를 CompanyRecruitments로 이름 변경
- 새로운 MyAccount 페이지 구현을 위해 기존 페이지의 역할 변경
